+++ apps/docs/game/render/Camera.tsx
@@ -7,12 +7,14 @@ import * as React from "react"
 
 export function CameraEntity() {
   return (
-    <ECS.Entity>
-      <ECS.Component name="isCamera" data={true} />
-      <ECS.Component name="sceneObject">
-        <PerspectiveCamera name="CameraEntity"></PerspectiveCamera>
-      </ECS.Component>
-    </ECS.Entity>
+    <>
+      <ECS.Entity>
+        <ECS.Component name="isCamera" data={true} />
+        <ECS.Component name="sceneObject">
+          <PerspectiveCamera name="CameraEntity"></PerspectiveCamera>
+        </ECS.Component>
+      </ECS.Entity>
+    </>
   )
 }