+++ packages/planets/src/math/Math.ts
@@ -26,6 +26,7 @@ export function randomSpherePoint(
   y0: number,
   z0: number,
   radius: number,
+  vector = new Vector3(),
 ) {
   let u = random()
   let v = random()
@@ -34,16 +35,7 @@ export function randomSpherePoint(
   let x = x0 + radius * Math.sin(phi) * Math.cos(theta)
   let y = y0 + radius * Math.sin(phi) * Math.sin(theta)
   let z = z0 + radius * Math.cos(phi)
-  return new Vector3(x, y, z)
-}
-
-function randomPointOnSphere(radius: number, origin: Vector3): Vector3 {
-  const angle1 = Math.random() * 2 * Math.PI
-  const angle2 = Math.random() * Math.PI
-  const x = radius * Math.sin(angle2) * Math.cos(angle1)
-  const y = radius * Math.sin(angle2) * Math.sin(angle1)
-  const z = radius * Math.cos(angle2)
-  return new Vector3(x, y, z).add(origin)
+  return vector.set(x, y, z)
 }
 
 export function randomSpherePointVector3(